requirements.txt
tweepy
pillow
matplotlib
